LPS-118700 Rename floating-toolbar directory to item-configuration-panels
