+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/Plugin.scss
@@ -1,5 +1,5 @@
-@import './floating-toolbar/ContainerStylesPanel';
-@import './floating-toolbar/RowStylesPanels';
+@import './item-configuration-panels/ContainerStylesPanel';
+@import './item-configuration-panels/RowStylesPanels';
 
 @import './ItemConfiguration';
 @import './PageStructureSidebar';
