+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/plugins/page-structure/components/item-configuration-panels/ContainerStylesPanel.test.js
@@ -19,7 +19,7 @@ import React from 'react';
 
 import {StoreAPIContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/store/index';
 import updateItemConfig from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig';
-import {ContainerStylesPanel} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/floating-toolbar/ContainerStylesPanel';
+import {ContainerStylesPanel} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/item-configuration-panels/ContainerStylesPanel';
 
 const renderComponent = ({itemConfig = {}, dispatch = () => {}} = {}) =>
 	render(
