+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/plugins/page-structure/components/item-configuration-panels/RowConfigurationPanel.test.js
@@ -17,11 +17,10 @@ import {cleanup, fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
 import {ResizeContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/components/ResizeContext';
-import {config} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/config/index';
 import {StoreAPIContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/store/index';
 import updateItemConfig from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig';
 import updateRowColumns from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateRowColumns';
-import {RowStylesPanel} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/floating-toolbar/RowStylesPanel';
+import {RowConfigurationPanel} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/item-configuration-panels/RowConfigurationPanel';
 
 const ITEM_CONFIG = {
 	gutters: true,
@@ -58,7 +57,7 @@ const renderComponent = ({
 			<ResizeContextProvider
 				value={{...RESIZE_CONTEXT_STATE, ...contextState}}
 			>
-				<RowStylesPanel
+				<RowConfigurationPanel
 					item={{
 						children: [],
 						config: {...ITEM_CONFIG, ...config},
@@ -79,7 +78,6 @@ jest.mock(
 				desktop: {label: 'Desktop'},
 				landscapeMobile: {label: 'landscapeMobile'},
 			},
-			responsiveEnabled: true,
 		},
 	})
 );
@@ -94,102 +92,37 @@ jest.mock(
 	() => jest.fn()
 );
 
-describe('RowStylesPanel', () => {
+describe('RowConfigurationPanel', () => {
 	afterEach(() => {
 		cleanup();
-		config.responsiveEnabled = true;
 		updateItemConfig.mockClear();
 		updateRowColumns.mockClear();
 	});
 
-	it('allows changing the modules per row', async () => {
+	it('allows changing the number of modules of a grid', async () => {
 		const {getByLabelText} = renderComponent({});
-		const input = getByLabelText('layout');
+		const input = getByLabelText('number-of-modules');
 
 		await fireEvent.change(input, {
-			target: {value: '2'},
+			target: {value: '6'},
 		});
 
-		expect(updateItemConfig).toHaveBeenCalledWith({
-			itemConfig: {
-				modulesPerRow: 2,
-			},
+		expect(updateRowColumns).toHaveBeenCalledWith({
 			itemId: '0',
+			numberOfColumns: 6,
 			segmentsExperienceId: '0',
+			viewportSizeId: 'desktop',
 		});
 	});
 
-	it('allows custom value in modules per row when row is customized', async () => {
-		const {getByLabelText} = renderComponent({
-			contextState: {customRow: true},
-		});
-		const input = getByLabelText('layout');
-
-		await fireEvent.change(input, {
-			target: {value: '0'},
-		});
-
-		expect(updateItemConfig).toHaveBeenCalledWith({
-			itemConfig: {
-				modulesPerRow: 'custom',
-			},
-			itemId: '0',
-			segmentsExperienceId: '0',
-		});
-	});
-
-	it('allows changing the vertical alignment', async () => {
+	it('allows changing the gutter', async () => {
 		const {getByLabelText} = renderComponent({});
-		const input = getByLabelText('vertical-alignment');
-
-		await fireEvent.change(input, {
-			target: {value: 'middle'},
-		});
-
-		expect(updateItemConfig).toHaveBeenCalledWith({
-			itemConfig: {
-				verticalAlignment: 'middle',
-			},
-			itemId: '0',
-			segmentsExperienceId: '0',
-		});
-	});
-
-	it('allows inverse order when number of modules is 2 and modules per row is 1', async () => {
-		const {getByLabelText} = renderComponent({
-			config: {
-				modulesPerRow: 1,
-			},
-		});
-		const input = getByLabelText('inverse-order');
+		const input = getByLabelText('show-gutter');
 
 		await fireEvent.click(input);
 
 		expect(updateItemConfig).toHaveBeenCalledWith({
-			itemConfig: {
-				reverseOrder: true,
-			},
-			itemId: '0',
-			segmentsExperienceId: '0',
-		});
-	});
-
-	it('allows changing configuration for a given viewport', async () => {
-		const {getByLabelText} = renderComponent({
-			state: {
-				selectedViewportSize: 'landscapeMobile',
-			},
-		});
-		const input = getByLabelText('layout');
-
-		await fireEvent.change(input, {
-			target: {value: '1'},
-		});
-
-		expect(updateItemConfig).toHaveBeenCalledWith({
-			itemConfig: {
-				landscapeMobile: {modulesPerRow: 1},
-			},
+			itemConfig: {gutters: false},
 			itemId: '0',
 			segmentsExperienceId: '0',
 		});
